LPS-59054 Move BundleUtil to BundleManager
